:hammer: [Refactoring] 컴파일 경고 없애기, 경고수준 4 설정, 경고를 오류로 처리
2026-07-22 회의 시 확인함

+++ LL2_Client_Win/CLogin.h
@@ -44,6 +44,6 @@ public:
 	virtual BOOL OnInitDialog();
 	void OnSocketConnect(BOOL bConnect);
 	int Login();
-	int OnLogin(const char *recvBuf, const int recvLen);
+	int OnLogin(const char *recvBuf, const size_t recvLen);
 
 };
